refact code
